Finished instrumentation code! Whohoo!
The next thing to do is to run it and see if it actually works. This can be done by using netcat to receive the UDP data and print it to stdout or dump it to a file.
Hey! That's a great idea! Use netcat to dump the instrumentation data to a file and then write a script/program to parse the file and generate meaningful statistics.