Changed field names to enable devices and content to be stored in the same Solr core
This code change must be used with an updated Solr schema that combines device and content records into one core. The solrdevurl and solrconurl must be set to point to the same core. As of right now, the devices_core has been updated to contain both device and content data and both of these parameters should be set to the pweb_devices core.